+++ src/main/groovy/fr/nereide/completion/provider/common/EntityFieldCompletionProvider.groovy
@@ -1,7 +1,22 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License") you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package fr.nereide.completion.provider.common
 
 import static com.intellij.psi.util.PsiTreeUtil.getChildrenOfType
-import static com.intellij.psi.util.PsiTreeUtil.getChildrenOfTypeAsList
 import static com.intellij.psi.util.PsiTreeUtil.getParentOfType
 import static fr.nereide.project.pattern.OfbizPluginConstants.DYNAMIC_VIEW_ENTITY_CLASS_NAME
 import static fr.nereide.project.worker.EntityWorker.getEntityFields
@@ -12,23 +27,17 @@ import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
 import com.intellij.codeInsight.completion.PrioritizedLookupElement
 import com.intellij.codeInsight.lookup.LookupElementBuilder
-import com.intellij.find.FindManager
-import com.intellij.find.findUsages.FindUsagesHandler
-import com.intellij.find.findUsages.FindUsagesOptions
-import com.intellij.find.impl.FindManagerImpl
-import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiType
 import com.intellij.psi.PsiVariable
 import com.intellij.usageView.UsageInfo
-import com.intellij.util.ArrayUtil
-import com.intellij.util.CommonProcessors
 import com.intellij.util.ProcessingContext
 import fr.nereide.dom.element.entitymodel.Entity
 import fr.nereide.dom.element.entitymodel.ViewEntity
 import fr.nereide.project.OfbizProjectHelper
 import fr.nereide.project.PluginActivator
 import fr.nereide.project.pattern.OfbizPluginConstants
+import fr.nereide.project.utils.PsiUtils
 import org.jetbrains.annotations.NotNull
 import org.jetbrains.plugins.groovy.lang.psi.api.statements.GrVariable
 import org.jetbrains.plugins.groovy.lang.psi.api.statements.expressions.GrReferenceExpression
@@ -72,27 +81,6 @@ abstract class EntityFieldCompletionProvider extends CompletionProvider<Completi
     }
 
     /**
-     * returns the list of all usages of the variable
-     * @param variable the psi variable that will be analyzed
-     * @return the usages of the variable
-     */
-    static List<UsageInfo> getUsagesOfVariable(PsiVariable variable) {
-        Project project = variable.project
-        FindUsagesHandler handler = ((FindManagerImpl) FindManager.getInstance(project)).findUsagesManager
-                .getFindUsagesHandler(variable, false)
-        if (!handler) return []
-        CommonProcessors.CollectProcessor<UsageInfo> processor =
-                new CommonProcessors.CollectProcessor<>(Collections.synchronizedList([]))
-        if (!processor) return null[]
-        List<PsiElement> psiElements = ArrayUtil.mergeArrays(handler.primaryElements, handler.secondaryElements)
-        FindUsagesOptions options = handler.getFindUsagesOptions(null)
-        for (PsiElement psiElement : psiElements) {
-            handler.processElementUsages(psiElement, processor, options)
-        }
-        return processor.results
-    }
-
-    /**
      * Searchs for the entity name in the query String
      */
     static String getEntityNameFromQuery(PsiElement element, Class methodTypeClass) {
@@ -121,7 +109,7 @@ abstract class EntityFieldCompletionProvider extends CompletionProvider<Completi
      * Tries to extract the entity name from the context and potentials assignements
      */
     String getEntityNameFromLastQueryAssignment(PsiVariable genericValueVariable) {
-        List<UsageInfo> usages = getUsagesOfVariable(genericValueVariable)
+        List<UsageInfo> usages = PsiUtils.getUsagesOfVariable(genericValueVariable)
         if (!usages) return ''
         UsageInfo lastQuery = usages.stream().filter { usage ->
             PsiElement assign = getParentOfType(usage.element, assigmentClass)
@@ -143,7 +131,7 @@ abstract class EntityFieldCompletionProvider extends CompletionProvider<Completi
         PsiElement[] params = getMethodArgs(addAliasInitialMethod)
         if (params) {
             String aliasToLookFor = params[index].text
-            List<UsageInfo> dveUsages = getUsagesOfVariable(initialDve)
+            List<UsageInfo> dveUsages = PsiUtils.getUsagesOfVariable(initialDve)
             PsiElement relevantAddAlias = dveUsages.collect { UsageInfo usage ->
                 getParentOfType(usage.element, methodExprClass)
             }.find { PsiElement addAliasCall ->
@@ -167,7 +155,7 @@ abstract class EntityFieldCompletionProvider extends CompletionProvider<Completi
 
     String getEntityNameFromInsideDynamisView(PsiElement element, Class methodClass) {
         PsiElement dveMethodCall = getParentOfType(element, methodClass)
-        PsiVariable dveVariable = getPsiTopVariable(dveMethodCall)
+        PsiVariable dveVariable = PsiUtils.getPsiTopVariable(dveMethodCall, referenceExpressionClass)
         if (dveVariable && variableHasDveType(dveVariable, dveMethodCall)) {
             return getEntityNameFromDynamicView(dveMethodCall, dveVariable)
         }
@@ -181,19 +169,8 @@ abstract class EntityFieldCompletionProvider extends CompletionProvider<Completi
         PsiElement dveMethodCall = getParentOfType(
                 getParentOfType(element, methodClass),
                 methodClass)
-        PsiVariable dveVariable = getPsiTopVariable(dveMethodCall)
+        PsiVariable dveVariable = PsiUtils.getPsiTopVariable(dveMethodCall, referenceExpressionClass)
         return dveVariable ? getEntityNameFromDynamicView(dveMethodCall, dveVariable, index) : null
-    }
-
-    /**
-     * Get the initial variable declaration
-     */
-    PsiVariable getPsiTopVariable(PsiElement fullCalledMethod) {
-        List fullGetStatementParts = getChildrenOfTypeAsList(fullCalledMethod, referenceExpressionClass)
-        if (!fullGetStatementParts) return null
-        List subGetStatementParts = getChildrenOfTypeAsList((fullGetStatementParts[0] as PsiElement),
-                referenceExpressionClass)
-        return subGetStatementParts ? subGetStatementParts[0].resolve() as PsiVariable : null
     }
 
     /**