+++ src/main/groovy/fr/nereide/completion/provider/java/JavaEntityFieldsFromGvMethodCompletionProvider.groovy
@@ -1,3 +1,19 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License") you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package fr.nereide.completion.provider.java
 
 import static com.intellij.psi.util.PsiTreeUtil.getParentOfType
@@ -7,6 +23,7 @@ import com.intellij.psi.PsiExpression
 import com.intellij.psi.PsiForeachStatement
 import com.intellij.psi.PsiMethodCallExpression
 import com.intellij.psi.PsiVariable
+import fr.nereide.project.utils.PsiUtils
 
 /**
  * Entity fields completion provider for java in dve
@@ -16,7 +33,7 @@ class JavaEntityFieldsFromGvMethodCompletionProvider extends JavaEntityFieldsCom
     @Override
     String getEntityNameFromPsiElement(PsiElement element) {
         PsiMethodCallExpression getMethod = getParentOfType(element, PsiMethodCallExpression)
-        PsiVariable gvVariable = getPsiTopVariable(getMethod)
+        PsiVariable gvVariable = PsiUtils.getPsiTopVariable(getMethod, referenceExpressionClass)
         if (!gvVariable) return null
         PsiExpression gvInit = gvVariable.initializer
         if (gvInit) {