+++ src/main/groovy/fr/nereide/project/worker/ServiceWorker.groovy
@@ -14,23 +14,35 @@ class ServiceWorker {
 
     final static String SERVICE_ATTR_NAME = 'attribute-name'
     final static String SERVICE_ATTR_TYPE = 'attribute-type'
-    public static final List<String> IN_SERVICE_MODE = ['IN', 'INOUT']
+    final static String INOUT = 'INOUT'
+    public static final List<String> IN_SERVICE_MODE = ['IN', INOUT]
+    public static final List<String> OUT_SERVICE_MODE = ['OUT', INOUT]
     public static final String TRUE = 'true'
     public static final String FALSE = 'false'
 
-    static List<Map> getRequiredInAttributes(Service service, OfbizProjectHelper ph) {
-        return getServiceInAttributes(service, ph, FALSE)
+    static List<Map> getRequiredInParameters(Service service, OfbizProjectHelper ph) {
+        return getServiceInParameters(service, ph, FALSE)
     }
 
-    static List<Map> getOptionalInAttributes(Service service, OfbizProjectHelper ph) {
-        return getServiceInAttributes(service, ph, TRUE)
+    static List<Map> getOptionalInParameters(Service service, OfbizProjectHelper ph) {
+        return getServiceInParameters(service, ph, TRUE)
     }
 
-    private static List<Map> getServiceInAttributes(Service service, OfbizProjectHelper ph, String optional) {
+    static List<Map> getServiceOutParameters(OfbizProjectHelper ph, String serviceName, String optional) {
+        return getServiceParameters(ph.getService(serviceName), ph, optional, OUT_SERVICE_MODE)
+    }
+
+    private static List<Map> getServiceInParameters(Service service, OfbizProjectHelper ph, String optional) {
+        return getServiceParameters(service, ph, optional, IN_SERVICE_MODE)
+    }
+
+    private static List<Map> getServiceParameters(Service service, OfbizProjectHelper ph, String optional, List modes) {
         List<ServiceAttribute> serviceAttributes = service.attributes
         List<Map> paramListToReturn = []
         // Vanilla params
-        paramListToReturn.addAll(collectParamsInService(serviceAttributes, optional))
+        paramListToReturn.addAll(collectParamsInService(serviceAttributes, optional, modes))
+        // Implements
+        paramListToReturn.addAll(getParamsFromServiceImplement(ph, service, optional, modes))
         // EntityAuto
         String defaultEntityName = service.defaultEntityName.value
         if (defaultEntityName && service.engine.value == 'entity-auto') {
@@ -52,12 +64,25 @@ class ServiceWorker {
         return paramListToReturn
     }
 
+    private static List<Map> getParamsFromServiceImplement(OfbizProjectHelper ph, Service service, String optional,
+                                                           List<String> modes) {
+        if (service.implements.size() == 0) {
+            return []
+        }
+        List<Service> implementedServices = service.implements.service*.value.collect {
+            serviceName -> ph.getService(serviceName)
+        }
+        return implementedServices.collect { Service implementedService ->
+            getServiceParameters(implementedService, ph, optional, modes)
+        }.flatten() as List<Map>
+    }
+
     private static List<LinkedHashMap<Object, Object>> collectParamsInService(List<ServiceAttribute> serviceAttributes,
-                                                                              String optional) {
-        return serviceAttributes.findAll { attr ->
-            optional == TRUE ? !isRequiredAttribute(attr) : isRequiredAttribute(attr) &&
-                    IN_SERVICE_MODE.contains(attr.mode.value)
-        }.collect { attr -> makeServiceAttrMap(attr) }
+                                                                              String optional, List<String> modes) {
+        return serviceAttributes
+                .findAll { attr -> modes.contains(attr.mode.value) }
+                .findAll { attr -> optional == TRUE ? !isRequiredAttribute(attr) : isRequiredAttribute(attr) }
+                .collect { attr -> makeServiceAttrMap(attr) }
     }
 
     private static boolean isEntityPk(List<String> pkFieldNames, EntityField it) {