+++ src/test/resources/testData/completion/groovy/ServiceOutParameterWithDsl.groovy
@@ -1,0 +1,2 @@
+Map result = run service: 'makeGame', with: [:]
+def foo = result.<caret>