+++ src/test/resources/testData/completion/groovy/ServiceOutParameterWithDslOnMapThirdUse.groovy
@@ -1,0 +1,6 @@
+Map result = [:]
+result = run service: 'makeHorse', with: [:]
+result = run service: 'makeWitcher', with: [:]
+result = run service: 'makeGame', with: [:]
+def foo = result.<caret>
+result = run service: 'makeHorse', with: [:]